OSIS-5779 pull members up
