+++ assessments/calendar/scores_exam_submission_calendar.py
@@ -24,8 +24,9 @@
 #
 ##############################################################################
 import datetime
+from typing import Optional
 
-from base.business.academic_calendar import AcademicEventSessionCalendarHelper
+from base.business.academic_calendar import AcademicEventSessionCalendarHelper, AcademicSessionEvent
 from base.models.academic_calendar import AcademicCalendar
 from base.models.academic_year import AcademicYear
 from base.models.enums.academic_calendar_type import AcademicCalendarTypes
@@ -34,6 +35,16 @@ from base.models.session_exam_calendar import SessionExamCalendar
 
 class ScoresExamSubmissionCalendar(AcademicEventSessionCalendarHelper):
     event_reference = AcademicCalendarTypes.SCORES_EXAM_SUBMISSION.name
+
+    def get_closest_academic_event(self, date=None) -> Optional[AcademicSessionEvent]:
+        opened_academic_events = self.get_opened_academic_events(date)
+        if opened_academic_events:
+            return opened_academic_events[0]
+
+        next_academic_event = self.get_next_academic_event(date)
+        if next_academic_event.session != self.FIRST_SESSION:
+            return next_academic_event
+        return None
 
     @classmethod
     def ensure_consistency_until_n_plus_6(cls):
